added date, handle expense cat limit

+++ src/main/java/com/example/financialtracker/report/PerMonthCatExpense.java
@@ -1,5 +1,6 @@
 package com.example.financialtracker.report;
 
+import jakarta.persistence.Tuple;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,14 +14,14 @@ import java.util.Map;
 @Getter
 @Setter
 public class PerMonthCatExpense {
-    private BigDecimal year;
-    private BigDecimal month;
+    private int year;
+    private int month;
     private BigDecimal total;
     private long count;
 
     public PerMonthCatExpense(Map<String, Object> report ) {
-        this.year = (BigDecimal) report.get("year");
-        this.month = (BigDecimal) report.get("month");
+        this.year = (int) report.get("year");
+        this.month = (int) report.get("month");
         this.total = (BigDecimal) report.get("total");
         this.count = (long) report.get("count");
     }
